refactor(markdown/*): migrate to .cjs
