+++ markdown/index.cjs
@@ -1,0 +1,7 @@
+const util = require("./utils/index.cjs");
+const m2h = require("./m2h/index.cjs");
+
+module.exports = {
+  ...util,
+  ...m2h,
+};
