--- markdown/index.js
@@ -1,2 +1,0 @@
-export * from "./utils/index.js";
-export * from "./m2h/index.js";
