+++ markdown/tsconfig.json
@@ -1,6 +1,7 @@
 {
   "extends": "../tsconfig.json",
   "compilerOptions": {
+    "module": "commonjs",
     "target": "es2019",
     "lib": ["esnext"],
     "skipLibCheck": true,
